+++ tests/testthat/test-regional_epinow.R
@@ -93,8 +93,6 @@ test_that("regional_epinow runs without error when given a very short timeout", 
 
 test_that("regional_epinow produces expected output when run with region specific settings", {
   skip_integration()
-  gp <- opts_list(gp_opts(), cases)
-  gp <- modifyList(gp, list(realland = NULL), keep.null = TRUE)
   rt <- opts_list(rt_opts(), cases, realland = rt_opts(rw = 7))
   delays <- opts_list(
     delay_opts(), cases, realland = delay_opts(example_reporting_delay)
@@ -104,7 +102,7 @@ test_that("regional_epinow produces expected output when run with region specifi
       data = cases,
       generation_time = generation_time_opts(example_generation_time),
       delays = delays,
-      rt = rt, gp = gp,
+      rt = rt,
       stan = stan_opts(
         samples = 100, warmup = 100,
         cores = 1, chains = 2,